prints prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	long long n;
-	unsigned int div = 2;
-	unsigned int ans = 0; 
-	unsigned int maxFact;
+	long n;
+	long div = 2;
+	long ans = 0; 
+	long maxFact;
 	n = 612852475143;
 	
 	while (n != 0)
